Refactored and put much sql stuff into top level class
